requirements.txt
python-telegram-bot==20.5
flask==2.2.5
requests==2.31.0